OggVorbisDecoder: fix integer comparison warnings in test case
Use size_t when comparing sample counts:

  test-oggvorbisdecoder.cpp: In function ‘void decodeWholeFile(const char*, int, int)’:
  test-oggvorbisdecoder.cpp:112:14: warning: comparison of integer expressions of different signedness: ‘size_t’ {aka ‘long unsigned int’} and ‘int’ [-Wsign-compare]
    112 |     assert(n == seconds * sampleRate);
        |            ~~^~~~~~~~~~~~~~~~~~~~~~~
  test-oggvorbisdecoder.cpp:114:24: warning: comparison of integer expressions of different signedness: ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
    114 |     assert(left.size() == n * sizeof(float));
        |            ~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~
  test-oggvorbisdecoder.cpp:115:25: warning: comparison of integer expressions of different signedness: ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
    115 |     assert(right.size() == n * sizeof(float));
        |            ~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~

+++ tests/test-oggvorbisdecoder.cpp
@@ -109,10 +109,10 @@ static void decodeWholeFile(const char *filename, int seconds, int sampleRate)
 
     size_t n = OggVorbisDecoder::decodeFile(filename, &left, &right,
                                             &actualSampleRate);
-    assert(n == seconds * sampleRate);
+    assert(n == static_cast<size_t>(seconds * sampleRate));
     assert(actualSampleRate == sampleRate);
-    assert(left.size() == n * sizeof(float));
-    assert(right.size() == n * sizeof(float));
+    assert(static_cast<size_t>(left.size()) == n * sizeof(float));
+    assert(static_cast<size_t>(right.size()) == n * sizeof(float));
 }
 
 static void testDecodeFileMono()
